refactor: 불필요한 커스텀 예외 제거
