+++ mvc/src/main/java/nextstep/mvc/adapter/AnnotationBasedAdapter.java
@@ -2,7 +2,6 @@ package nextstep.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.adapter.HandlerAdapter;
 import nextstep.mvc.handler.annotation.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
