+++ mvc/src/main/java/nextstep/mvc/adapter/HandlerAdapters.java
@@ -1,5 +1,6 @@
 package nextstep.mvc.adapter;
 
+import nextstep.mvc.exception.AdapterNotFoundException;
 import nextstep.mvc.handler.controller.ControllerBasedAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
