+++ mvc/src/main/java/nextstep/mvc/exception/AdapterNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.adapter;
+package nextstep.mvc.exception;
 
 public class AdapterNotFoundException extends RuntimeException {
