+++ mvc/src/main/java/nextstep/mvc/exception/MvcException.java
@@ -1,6 +1,10 @@
-package nextstep.mvc.view;
+package nextstep.mvc.exception;
 
 public class MvcException extends RuntimeException {
+
+    public MvcException(String message) {
+        super(message);
+    }
 
     public MvcException(String message, Throwable cause) {
         super(message, cause);
