+++ mvc/src/main/java/nextstep/mvc/view/JsonView.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.exception.MvcException;
 import nextstep.web.support.MediaType;
 
 import java.io.IOException;
